Update testimonial link so it stays valid
Hi! I'm @scatmandan on Twitter. Along with most of the world, I'm imminently about to delete my Twitter account, which will break the testimonial link on your homepage (I've deleted most of my tweets already, but I've left that one up for now).

The tweet in which I gave you that testimonial was backed up to my blog, though: it can be found intact at https://danq.me/2021/09/29/freshrss-addiction/. So if you want that link to stay valid, you might like to update it to that URL instead! This PR proposes the necessary change.

(If you don't want to make that change, or you'd rather seek a different testimonial, that's fine too. Either way, I'll remain a huge FreshRSS fan. Congratulations on 10 years!)