Updated the cloud providers.
Removed:
1. vern instance as is dead for sometime now, and the whole vern domain itself is dead.
2. Facil services, as they no longer provide for FreshRSS instance.
3. Zaclys instance as they too no longer provide RSS instance.

Updated the RSS direct instance details and URLs, so that they take users to the correct place instead of having to find, navigate, locate and what not in the pages of the domains they are hosted in. Makes it much simpler and directs to path of the site itself.